wallet plugin port from other branch
