+++ lib/wallet/http.js
@@ -1250,38 +1250,38 @@ class HTTP extends Server {
     });
 
     // Batch Open
-    this.post('/wallet/:id/batch/open', async (req, res) => {
-      const valid = Validator.fromRequest(req);
-      const names = valid.array('names');
-      const force = valid.bool('force', false);
-      const passphrase = valid.str('passphrase');
-      const broadcast = valid.bool('broadcast', true);
-      const sign = valid.bool('sign', true);
-      const MAX_NAME_ARRAY_LENGTH = 200;
+    // this.post('/wallet/:id/batch/open', async (req, res) => {
+    //   const valid = Validator.fromRequest(req);
+    //   const names = valid.array('names');
+    //   const force = valid.bool('force', false);
+    //   const passphrase = valid.str('passphrase');
+    //   const broadcast = valid.bool('broadcast', true);
+    //   const sign = valid.bool('sign', true);
+    //   const MAX_NAME_ARRAY_LENGTH = 200;
 
-      assert(names && names.length > 0, 'Names are required.');
-      assert(names.length <= MAX_NAME_ARRAY_LENGTH,
-        `Names array shoud not exceed ${MAX_NAME_ARRAY_LENGTH}`);
-      assert(broadcast ? sign : true, 'Must sign when broadcasting.');
+    //   assert(names && names.length > 0, 'Names are required.');
+    //   assert(names.length <= MAX_NAME_ARRAY_LENGTH,
+    //     `Names array shoud not exceed ${MAX_NAME_ARRAY_LENGTH}`);
+    //   assert(broadcast ? sign : true, 'Must sign when broadcasting.');
 
-      const options = TransactionOptions.fromValidator(valid);
-      const {mtx, errors, isAllError} = await req.wallet
-                                        .createBatchOpen(names, force, options);
+    //   const options = TransactionOptions.fromValidator(valid);
+    //   const {mtx, errors, isAllError} = await req.wallet
+    //                                     .createBatchOpen(names, force, options);
 
-      if (isAllError) // no valid output in mtx
-        return res.json(500, {errors: errors});
+    //   if (isAllError) // no valid output in mtx
+    //     return res.json(500, {errors: errors});
 
-      if (broadcast) {
-        const tx = await req.wallet.sendMTX(mtx, passphrase);
-        return res.json(200, {tx: tx.getJSON(this.network), errors: errors});
-      }
+    //   if (broadcast) {
+    //     const tx = await req.wallet.sendMTX(mtx, passphrase);
+    //     return res.json(200, {tx: tx.getJSON(this.network), errors: errors});
+    //   }
 
-      if (sign) {
-        await req.wallet.sign(mtx, passphrase);
-      }
+    //   if (sign) {
+    //     await req.wallet.sign(mtx, passphrase);
+    //   }
 
-      return res.json(200, {tx: mtx.getJSON(this.network), errors: errors});
-    });
+    //   return res.json(200, {tx: mtx.getJSON(this.network), errors: errors});
+    // });
 
     // Create Bid
     this.post('/wallet/:id/bid', async (req, res) => {
@@ -2170,16 +2170,23 @@ class HTTP extends Server {
    * @param {*} path location handler is listening
    * @param {*} handler http handler that will handle requests
    * arriving to `path` registered
+   * @param {Object} context context
    * @returns {undefined}
    */
-  registerRoute(httpMethod, path, handler) {
-    const validHttpMethods = [ 'post', 'get', 'put', 'patch', 'delete'];
-    assert(httpMethod && typeof httpMethod === 'string');
-    assert(validHttpMethods.includes(httpMethod.toLowerCase()));
+  registerRoute(httpMethod, path, handler, context) {
+    const validHttpMethods = ['post', 'get', 'put', 'patch', 'delete'];
+    assert(httpMethod && typeof httpMethod === 'string',
+    'httpMethod parameter must be provided and it must be a string');
+    assert(validHttpMethods.includes(httpMethod.toLowerCase()),
+    'httpMethod must be one of ' + validHttpMethods.join(','));
+
+    // update context
+    context.TransactionOptions = TransactionOptions;
+    context.Validator = Validator;
+    context.network = this.network;
 
     this[httpMethod](path, async (req, res) => {
-      const context = {TransactionOptions, Validator};
-      handler(req, res, context);
+      await handler(req, res, context);
     });
   }
 }
