+++ lib/wallet/plugin.js
@@ -12,6 +12,9 @@ const NodeClient = require('./nodeclient');
 const HTTP = require('./http');
 const RPC = require('./rpc');
 
+// TODO Remove me
+const assert = require('bsert');
+
 /**
  * @exports wallet/plugin
  */
@@ -74,7 +77,49 @@ class Plugin extends EventEmitter {
       adminToken: this.config.str('admin-token')
     });
 
+    // Register Additional endpoints
+    const context = {};
+    this.http.registerRoute('post','/wallet/:id/batch/open',
+      this.handleBatchOpen,
+      context);
+
     this.init();
+  }
+
+  async handleBatchOpen(req, res, context) {
+    // Load Helper classes from context
+    const {TransactionOptions,Validator,network} = context;
+
+    const valid = Validator.fromRequest(req);
+    const names = valid.array('names');
+    const force = valid.bool('force', false);
+    const passphrase = valid.str('passphrase');
+    const broadcast = valid.bool('broadcast', true);
+    const sign = valid.bool('sign', true);
+    const MAX_NAME_ARRAY_LENGTH = 200;
+
+    assert(names && names.length > 0, 'Names are required.');
+    assert(names.length <= MAX_NAME_ARRAY_LENGTH,
+      `Names array shoud not exceed ${MAX_NAME_ARRAY_LENGTH}`);
+    assert(broadcast ? sign : true, 'Must sign when broadcasting.');
+
+    const options = TransactionOptions.fromValidator(valid);
+    const {mtx, errors, isAllError} = await req.wallet
+                                      .createBatchOpen(names, force, options);
+
+    if (isAllError) // no valid output in mtx
+      return res.json(500, {errors: errors});
+
+    if (broadcast) {
+      const tx = await req.wallet.sendMTX(mtx, passphrase);
+      return res.json(200, {tx: tx.getJSON(network), errors: errors});
+    }
+
+    if (sign) {
+      await req.wallet.sign(mtx, passphrase);
+    }
+
+    return res.json(200, {tx: mtx.getJSON(network), errors: errors});
   }
 
   init() {
