+++ package.json
@@ -46,8 +46,9 @@
     "bweb": "~0.1.10",
     "goosig": "~0.10.0",
     "hs-client": "~0.0.9",
-    "n64": "~0.2.10",
+    "hsd-plugin-template": "^1.0.2",
     "mrmr": "~0.1.8",
+    "n64": "~0.2.10",
     "namebase-hs-client": "github:namebasehq/hs-client#799eb9b",
     "urkel": "~0.7.0"
   },
